models are divided between apps and admin.py created for each ot them
