--- app/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-from cargosystem.models import User, Currency, Wallet, Country
-
-admin.site.register(User)
-admin.site.register(Currency)
-admin.site.register(Wallet)
-admin.site.register(Country)
-
