+++ core/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from core.models import Country
+
+
+admin.site.register(Country)
+
