+++ core/models.py
@@ -1,0 +1,8 @@
+from django.db import models
+
+
+
+class Country(models.Model):
+    country_name = models.CharField(max_length=100)
+    row_status = models.IntegerField(default = 1)
+
