+++ docker-compose.yml
@@ -9,6 +9,7 @@ services:
       - POSTGRES_DB=postgres
       - POSTGRES_USER=postgres
       - POSTGRES_PASSWORD=postgres
+
   web:
     build: .
     command: python manage.py runserver 0.0.0.0:8000
@@ -20,6 +21,16 @@ services:
       - POSTGRES_NAME=postgres
       - POSTGRES_USER=postgres
       - POSTGRES_PASSWORD=postgres
-      - POSTGRES_HOST=db  # This line is new
+      - POSTGRES_HOST=db
+    depends_on:
+      - db
+
+  pgadmin:
+    image: dpage/pgadmin4
+    environment:
+      - PGADMIN_DEFAULT_EMAIL=vnasibov7929@gmail.com
+      - PGADMIN_DEFAULT_PASSWORD=Salam123!
+    ports:
+      - "8080:80"
     depends_on:
       - db
