+++ user/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from user.models import User
+
+admin.site.register(User)
+
+
