+++ user/models.py
@@ -45,7 +45,7 @@ class User(models.Model):
     (4, "DYI"),
 ]
 
-    gov_prefix_id = models.IntegerField(max_length=1, choices=GOV_PREFIX_CHOICES)
+    gov_prefix_id = models.IntegerField(validators=[MinLengthValidator(1), MaxLengthValidator(4)], choices=GOV_PREFIX_CHOICES)
     gov_id = models.IntegerField(validators=[
         MinValueValidator(1),
         MaxValueValidator(9999999)  # maximum value for 7 digits
@@ -69,19 +69,4 @@ class User(models.Model):
 
     selected_local_warehouse_id = models.IntegerField()
 
-
-class Currency(models.Model):
-    currency_name = models.CharField(max_length=5)
-    currency_sign = models.CharField(max_length=1)
-    currency_rate = models.FloatField()
-
-
-class Wallet(models.Model):
-    user_id = models.ForeignKey(User, on_delete=models.CASCADE)
-    currency_id = models.ForeignKey(Currency, on_delete=models.CASCADE)
-    balance = models.DecimalField(max_digits=5, decimal_places=2)
-
-class Country(models.Model):
-    country_name = models.CharField(max_length=100)
-    row_status = models.IntegerField(default = 1)
